feature: 프로필 수정 및 컴포넌트 분리
