+++ src/utils/login.js
@@ -1,0 +1,4 @@
+export const isLoggedIn = () => {
+  const user = localStorage.getItem("user");
+  return user ? true : false;
+};
