+++ src/utils/staticData.js
@@ -1,0 +1,32 @@
+export const postData = [
+  {
+    id: 0,
+    username: "홍길동",
+    content: "오늘 날씨가 정말 좋네요. 다들 좋은 하루 보내세요!",
+    createdAt: "5분전",
+  },
+  {
+    id: 1,
+    username: "김철수",
+    content: "새로운 프로젝트를 시작했어요. 열심히 코딩 중입니다!",
+    createdAt: "15분전",
+  },
+  {
+    id: 2,
+    username: "이영희",
+    content: "오늘 점심 메뉴 추천 받습니다. 뭐가 좋을까요?",
+    createdAt: "30분전",
+  },
+  {
+    id: 3,
+    username: "박민수",
+    content: "주말에 등산 가실 분 계신가요? 함께 가요!",
+    createdAt: "1시간전",
+  },
+  {
+    id: 4,
+    username: "정수연",
+    content: "새로 나온 영화 재미있대요. 같이 보러 갈 사람?",
+    createdAt: "2시간전",
+  },
+];
